#401. Refactor - remove RawSourceJudgmentParser
