+++ saos-batch/src/main/java/pl/edu/icm/saos/batch/core/CtjImportJobConfiguration.java
@@ -12,7 +12,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import pl.edu.icm.saos.importer.common.JudgmentImportDownloadWriter;
-import pl.edu.icm.saos.importer.notapi.common.StringItemImportDownloadProcessor;
+import pl.edu.icm.saos.importer.notapi.common.JsonImportDownloadProcessor;
 import pl.edu.icm.saos.importer.notapi.common.JsonImportDownloadReader;
 import pl.edu.icm.saos.importer.notapi.common.NotApiImportDownloadStepExecutionListener;
 import pl.edu.icm.saos.persistence.model.importer.notapi.RawSourceCtJudgment;
@@ -36,7 +36,7 @@ public class CtjImportJobConfiguration {
     
     @Autowired
     @Qualifier("ctjImportDownloadProcessor")
-    private StringItemImportDownloadProcessor<RawSourceCtJudgment> ctjImportDownloadProcessor;
+    private JsonImportDownloadProcessor<RawSourceCtJudgment> ctjImportDownloadProcessor;
     
     @Autowired
     private JudgmentImportDownloadWriter ctjImportDownloadWriter;
