+++ saos-import/src/main/java/pl/edu/icm/saos/importer/notapi/supremecourt/SupremeCourtImportConfiguration.java
@@ -12,12 +12,11 @@ import pl.edu.icm.saos.importer.common.converter.JudgmentConverter;
 import pl.edu.icm.saos.importer.common.converter.JudgmentConverterImpl;
 import pl.edu.icm.saos.importer.common.overwriter.DelegatingJudgmentOverwriter;
 import pl.edu.icm.saos.importer.common.overwriter.JudgmentOverwriter;
-import pl.edu.icm.saos.importer.notapi.common.JudgmentImportProcessReader;
-import pl.edu.icm.saos.importer.notapi.common.JsonJudgmentImportProcessProcessor;
-import pl.edu.icm.saos.importer.notapi.common.StringItemImportDownloadProcessor;
 import pl.edu.icm.saos.importer.notapi.common.JsonImportDownloadReader;
+import pl.edu.icm.saos.importer.notapi.common.JsonJudgmentImportProcessProcessor;
+import pl.edu.icm.saos.importer.notapi.common.JudgmentImportProcessReader;
 import pl.edu.icm.saos.importer.notapi.common.NotApiImportDownloadStepExecutionListener;
-import pl.edu.icm.saos.importer.notapi.supremecourt.judgment.download.RawSourceScJudgmentParser;
+import pl.edu.icm.saos.importer.notapi.common.JsonImportDownloadProcessor;
 import pl.edu.icm.saos.importer.notapi.supremecourt.judgment.json.SourceScJudgment;
 import pl.edu.icm.saos.importer.notapi.supremecourt.judgment.process.SourceScJudgmentExtractor;
 import pl.edu.icm.saos.persistence.model.SupremeCourtJudgment;
@@ -42,9 +41,6 @@ public class SupremeCourtImportConfiguration {
     
     @Autowired
     private MappingJsonFactory jsonFactory;
-    
-    @Autowired
-    private RawSourceScJudgmentParser rawSourceScJudgmentParser;
     
     
     @Autowired 
@@ -77,9 +73,9 @@ public class SupremeCourtImportConfiguration {
     }
     
     @Bean
-    public StringItemImportDownloadProcessor<RawSourceScJudgment> scjImportDownloadProcessor() {
-        StringItemImportDownloadProcessor<RawSourceScJudgment> scjImportDownloadProcessor = new StringItemImportDownloadProcessor<RawSourceScJudgment>();
-        scjImportDownloadProcessor.setRawSourceJudgmentParser(rawSourceScJudgmentParser);
+    public JsonImportDownloadProcessor<RawSourceScJudgment> scjImportDownloadProcessor() {
+        JsonImportDownloadProcessor<RawSourceScJudgment> scjImportDownloadProcessor = new JsonImportDownloadProcessor<>(RawSourceScJudgment.class);
+        scjImportDownloadProcessor.setSourceJudgmentParser(sourceScJudgmentParser());
         
         return scjImportDownloadProcessor;
     }
