+++ saos-import/src/main/java/pl/edu/icm/saos/importer/notapi/supremecourt/judgment/json/SourceScJudgment.java
@@ -2,17 +2,11 @@ package pl.edu.icm.saos.importer.notapi.supremecourt.judgment.json;
 
 import java.util.List;
 
-import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
-import org.joda.time.DateTime;
-import org.joda.time.LocalDate;
-
-import pl.edu.icm.saos.importer.notapi.common.DateTimeDeserializer;
-import pl.edu.icm.saos.importer.notapi.common.LocalDateIsoDeserializer;
+import pl.edu.icm.saos.importer.notapi.common.SourceJudgment;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.common.collect.Lists;
 
 /**
@@ -21,37 +15,20 @@ import com.google.common.collect.Lists;
  * @author Łukasz Dumiszewski
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class SourceScJudgment {
+public class SourceScJudgment extends SourceJudgment {
     
-    private String textContent;
-    private Source source;
     private String caseNumber;
-    private LocalDate judgmentDate;
     private String supremeCourtJudgmentForm;
     private String personnelType;
     private List<String> supremeCourtChambers = Lists.newArrayList();
     private String supremeCourtChamberDivision;
-    private List<SourceScJudge> judges = Lists.newArrayList();
+    
     
     //------------------------ GETTERS --------------------------
-
-    public String getTextContent() {
-        return textContent;
-    }
-
-    @NotNull
-    public Source getSource() {
-        return source;
-    }
 
     @NotNull
     public String getCaseNumber() {
         return caseNumber;
-    }
-    
-    @JsonDeserialize(using = LocalDateIsoDeserializer.class)
-    public LocalDate getJudgmentDate() {
-        return judgmentDate;
     }
 
     public String getSupremeCourtJudgmentForm() {
@@ -70,10 +47,7 @@ public class SourceScJudgment {
         return supremeCourtChamberDivision;
     }
 
-    @Valid
-    public List<SourceScJudge> getJudges() {
-        return judges;
-    }
+
 
 
 
@@ -81,20 +55,8 @@ public class SourceScJudgment {
     
     //------------------------ SETTERS --------------------------
     
-    public void setSource(Source source) {
-        this.source = source;
-    }
-
-    public void setTextContent(String textContent) {
-        this.textContent = textContent;
-    }
-    
     public void setCaseNumber(String caseNumber) {
         this.caseNumber = caseNumber;
-    }
-
-    public void setJudgmentDate(LocalDate judgmentDate) {
-        this.judgmentDate = judgmentDate;
     }
 
     public void setSupremeCourtJudgmentForm(String supremeCourtJudgmentForm) {
@@ -113,106 +75,5 @@ public class SourceScJudgment {
         this.supremeCourtChamberDivision = supremeCourtChamberDivision;
     }
 
-    public void setJudges(List<SourceScJudge> judges) {
-        this.judges = judges;
-    }
 
-
- 
-
-
-
-    
-    
-    
-    
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public static class Source {
-        
-        private String sourceJudmentId;
-        private String sourceJudgmentUrl;
-        private DateTime publicationDateTime;
-        
-        
-        //------------------------ GETTERS --------------------------
-        
-        @NotNull
-        public String getSourceJudgmentId() {
-            return sourceJudmentId;
-        }
-        
-        @NotNull
-        public String getSourceJudgmentUrl() {
-            return sourceJudgmentUrl;
-        }
-        
-        @JsonDeserialize(using = DateTimeDeserializer.class)
-        public DateTime getPublicationDateTime() {
-            return publicationDateTime;
-        }
-        
-        
-        //------------------------ SETTERS --------------------------
-        
-        public void setSourceJudgmentId(String sourceJudgmentId) {
-            this.sourceJudmentId = sourceJudgmentId;
-        }
-        public void setSourceJudgmentUrl(String sourceJudgmentUrl) {
-            this.sourceJudgmentUrl = sourceJudgmentUrl;
-        }
-        public void setPublicationDateTime(DateTime publicationDateTime) {
-            this.publicationDateTime = publicationDateTime;
-        }
-       
-    }
-
-
-
-
-
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public static class SourceScJudge {
-        
-        private String name;
-        private String function;
-        private List<String> specialRoles = Lists.newArrayList();
-        
-        
-        //------------------------ GETTERS --------------------------
-        
-        @NotNull
-        public String getName() {
-            return name;
-        }
-        public String getFunction() {
-            return function;
-        }
-        public List<String> getSpecialRoles() {
-            return specialRoles;
-        }
-        
-        
-        //------------------------ SETTERS --------------------------
-        
-        public void setName(String name) {
-            this.name = name;
-        }
-        public void setFunction(String function) {
-            this.function = function;
-        }
-        public void setSpecialRoles(List<String> specialRoles) {
-            this.specialRoles = specialRoles;
-        }
-        
-        
-        
-        
-    }
-
-
-
-
- 
-
-    
 }
